+++ src/lib/bits/date-range-picker/components/date-range-picker-day.svelte
@@ -1,9 +1,9 @@
 <script lang="ts">
 	import { melt } from "@melt-ui/svelte";
 	import { getCtx, getCalendarAttrs } from "../ctx.js";
-	import type { DateProps } from "../types.js";
+	import type { DayProps } from "../types.js";
 
-	type $$Props = DateProps;
+	type $$Props = DayProps;
 
 	export let date: $$Props["date"];
 	export let month: $$Props["month"];
@@ -14,23 +14,19 @@
 		helpers: { isDateDisabled, isDateUnavailable }
 	} = getCtx();
 
-	const attrs = getCalendarAttrs("date");
+	const attrs = getCalendarAttrs("day");
 
 	$: builder = $cell(date, month);
 	$: Object.assign(builder, attrs);
-
-	$: slotProps = {
-		builder,
-		disabled: $isDateDisabled(date),
-		unavailable: $isDateUnavailable(date)
-	};
+	$: disabled = $isDateDisabled(date);
+	$: unavailable = $isDateUnavailable(date);
 </script>
 
 {#if asChild}
-	<slot {...slotProps} />
+	<slot {builder} {disabled} {unavailable} />
 {:else}
 	<div use:melt={builder} {...$$restProps}>
-		<slot {...slotProps}>
+		<slot {builder} {disabled} {unavailable}>
 			{date.day}
 		</slot>
 	</div>