+++ src/lib/bits/index.ts
@@ -28,4 +28,5 @@ export * as Switch from "./switch/index.js";
 export * as Tabs from "./tabs/index.js";
 export * as Toggle from "./toggle/index.js";
 export * as ToggleGroup from "./toggle-group/index.js";
+export * as Toolbar from "./toolbar/index.js";
 export * as Tooltip from "./tooltip/index.js";