+++ src/lib/bits/range-calendar/_types.ts
@@ -48,6 +48,7 @@ type Props = Expand<
 		 * A callback function called when the placeholder changes.
 		 */
 		onPlaceholderChange?: OnChangeFn<DateValue>;
+
 		/**
 		 * If `true`, the calendar will focus the selected day,
 		 * today, or the first day of the month in that order depending
@@ -56,6 +57,24 @@ type Props = Expand<
 		 * @default false
 		 */
 		initialFocus?: boolean;
+
+		/**
+		 * The `start` value of the date range, which can exist prior
+		 * to the `value` being set. The `value` is only set once a `start`
+		 * and `end` value are selected.
+		 *
+		 * You can `bind:startValue` to a value to receive updates outside
+		 * this component when the user selects a `start` value.
+		 *
+		 * Modifying this value outside the component will have no effect.
+		 * To programmatically control the `start` value, use `bind:value`
+		 * and update the `start` property of the `DateRange` object.
+		 *
+		 * This is provided as a convenience for use cases where you want
+		 * to display the selected `start` value outside the component before
+		 * the `value` is set.
+		 */
+		startValue?: DateValue | undefined;
 	} & AsChild
 >;
 
@@ -77,7 +96,7 @@ type GridBodyProps = AsChild;
 
 type GridRowProps = AsChild;
 
-type BaseDateProps = Expand<
+type BaseDayProps = Expand<
 	{
 		/**
 		 * The date value of the cell.
@@ -91,9 +110,9 @@ type BaseDateProps = Expand<
 	} & AsChild
 >;
 
-type CellProps = Expand<Omit<BaseDateProps, "month">>;
+type CellProps = Expand<Omit<BaseDayProps, "month">>;
 
-type DateProps = Expand<BaseDateProps>;
+type DayProps = Expand<BaseDayProps>;
 
 export type {
 	Props,
@@ -107,5 +126,5 @@ export type {
 	GridBodyProps,
 	GridRowProps,
 	CellProps,
-	DateProps
+	DayProps
 };