+++ src/lib/bits/toolbar/_export.types.ts
@@ -1,0 +1,11 @@
+export type {
+	Props as ToolbarProps,
+	ButtonProps as ToolbarButtonProps,
+	LinkProps as ToolbarLinkProps,
+	GroupProps as ToolbarGroupProps,
+	GroupItemProps as ToolbarItemProps,
+	//
+	ButtonEvents as ToolbarButtonEvents,
+	LinkEvents as ToolbarLinkEvents,
+	GroupItemEvents as ToolbarItemEvents
+} from "./types.js";